getting right of last warnings

+++ urbanAlphabetsV 3/Nav/views/AddAlphabet.m
@@ -136,7 +136,7 @@
     float firstYPos=firstShapeY;
     yPos=yPos-firstYPos;
     float elementNumber=yPos/clickedObject.frame.size.height;
-    elementNoChosen=lroundf(elementNumber);
+    elementNoChosen=(int)lroundf(elementNumber);
     for (int i=0; i<[shapesForBackground count]; i++) {
         UIView *shape=[shapesForBackground objectAtIndex:i];
         if (i==elementNoChosen) {
